// backend/src/modules/task/task.enums.ts
export enum TaskType {

    READING = 'READING',

    VIDEO = 'VIDEO',

    QUIZ = 'QUIZ',

    CODING = 'CODING',

    PROJECT = 'PROJECT',

    PRACTICE = 'PRACTICE',

    ASSIGNMENT = 'ASSIGNMENT'
}

export enum CompletionType {

    SELF = 'SELF',

    QUIZ = 'QUIZ',

    AI_REVIEW = 'AI_REVIEW',

    MANUAL = 'MANUAL'
}